third_party/WebKit/LayoutTests/svg/animations/svg-animation-policy-once-expected.txt

SVG with animation policy, once

This tests svg animation with animation policy once

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.animationsPaused() is false
PASS rootSVGElement.animationsPaused() is false
PASS rootSVGElement.getCurrentTime() is 2.5
PASS rect.y.animVal.value is 100
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.animationsPaused() is false
PASS rootSVGElement.getCurrentTime() is 5.0
PASS rect.y.animVal.value is 100
PASS rootSVGElement.animationsPaused() is false
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.animationsPaused() is true
PASS rootSVGElement.animationsPaused() is false
PASS rootSVGElement.getCurrentTime() is 5.0
PASS rect.y.animVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
